+++ packages/sources/coinmetrics/src/index.ts
@@ -1,13 +1,13 @@
 import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
 import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { config } from './config'
-import { burned, lwba, price, totalBurned } from './endpoint'
+import { burned, lwba, price, totalBurned, realizedVol } from './endpoint'
 
 export const adapter = new PriceAdapter({
   defaultEndpoint: price.name,
   name: 'COINMETRICS',
   config,
-  endpoints: [price, lwba, totalBurned, burned],
+  endpoints: [price, lwba, totalBurned, burned, realizedVol],
   rateLimiting: {
     tiers: {
       community: {