+++ packages/sources/finnhub/src/endpoint/quote.ts
@@ -13,25 +13,33 @@ import { makeLogger } from '@chainlink/external-adapter-framework/util'
 
 const logger = makeLogger('Finnhub Quote')
 
-export const inputParameters = new InputParameters({
-  base: {
-    aliases: ['from', 'coin'],
-    type: 'string',
-    description: 'The symbol of symbols of the currency to query',
-    required: true,
+export const inputParameters = new InputParameters(
+  {
+    base: {
+      aliases: ['from', 'coin'],
+      type: 'string',
+      description: 'The symbol of symbols of the currency to query',
+      required: true,
+    },
+    quote: {
+      aliases: ['to', 'market'],
+      type: 'string',
+      description: 'The symbol of the currency to convert to',
+      required: false,
+    },
+    exchange: {
+      type: 'string',
+      description: 'The exchange to fetch data for',
+      required: false,
+    },
   },
-  quote: {
-    aliases: ['to', 'market'],
-    type: 'string',
-    description: 'The symbol of the currency to convert to',
-    required: false,
-  },
-  exchange: {
-    type: 'string',
-    description: 'The exchange to fetch data for',
-    required: false,
-  },
-})
+  [
+    {
+      base: 'EUR',
+      quote: 'USD',
+    },
+  ],
+)
 
 export type BaseEndpointTypes = {
   Parameters: typeof inputParameters.definition
@@ -130,7 +138,15 @@ const validateExchange = (req: AdapterRequest<typeof inputParameters.validated>)
 export const buildQuoteEndpoint = (overrides?: Record<string, string>) =>
   new PriceEndpoint<BaseEndpointTypes>({
     name: 'quote',
-    aliases: ['common', 'stock', 'forex'],
+    aliases: [
+      'common',
+      'stock',
+      'forex',
+      'commodities',
+      'forex-quote',
+      'equity-quote',
+      'commodity-quote',
+    ],
     transportRoutes: new TransportRoutes<BaseEndpointTypes>()
       .register('ws', wsTransport)
       .register('rest', httpTransport),