+++ packages/sources/kaiko/src/index.ts
@@ -2,12 +2,12 @@ import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
 import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { config } from './config'
 import includes from './config/includes.json'
-import { trades } from './endpoint'
+import { trades, realizedVolatility } from './endpoint'
 
 export const adapter = new PriceAdapter({
   defaultEndpoint: trades.name,
   name: 'KAIKO',
-  endpoints: [trades],
+  endpoints: [trades, realizedVolatility],
   config,
   includes,
   rateLimiting: {