+++ packages/sources/layer2-sequencer-health/test/unit/health.test.ts
@@ -111,7 +111,7 @@ describe('adapter', () => {
       expect(response.data.result).toBe(1)
     })
 
-    it('Empty transaction check has the final word on unhealthy method responses', async () => {
+    it('Empty transaction check has the final word on unhealthy method responses if requireTxFailure = true', async () => {
       jest.spyOn(network, 'checkSequencerHealth').mockReturnValue(Promise.resolve(false))
       jest.spyOn(network, 'checkNetworkProgress').mockReturnValue(Promise.resolve(false))
       jest.spyOn(network, 'getStatusByTransaction').mockReturnValue(Promise.resolve(true))
@@ -120,12 +120,31 @@ describe('adapter', () => {
         {
           data: {
             network: 'arbitrum',
+            requireTxFailure: true,
           },
         } as AdapterRequest<TInputParameters>,
         {},
       )
 
       expect(response.data.result).toBe(0)
+    })
+
+    it('Empty transaction check does not impact response if requireTxFailure = false (default)', async () => {
+      jest.spyOn(network, 'checkSequencerHealth').mockReturnValue(Promise.resolve(false))
+      jest.spyOn(network, 'checkNetworkProgress').mockReturnValue(Promise.resolve(false))
+      jest.spyOn(network, 'getStatusByTransaction').mockReturnValue(Promise.resolve(true))
+
+      const response = await execute(
+        {
+          data: {
+            network: 'arbitrum',
+            requireTxFailure: false,
+          },
+        } as AdapterRequest<TInputParameters>,
+        {},
+      )
+
+      expect(response.data.result).toBe(1)
     })
 
     it('Empty transaction confirms rest of methods', async () => {