+++ packages/composites/historical-average/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @chainlink/historical-average-adapter
 
+## 1.2.50
+
+### Patch Changes
+
+- Updated dependencies [[`112d9c967`](https://github.com/smartcontractkit/external-adapters-js/commit/112d9c9679fdcb3210ec53a17b6b8f0016ad0a0d)]:
+  - @chainlink/coinmarketcap-adapter@2.0.17
+
+## 1.2.49
+
+### Patch Changes
+
+- Updated dependencies [[`a7c807fc6`](https://github.com/smartcontractkit/external-adapters-js/commit/a7c807fc6ce96059c1324381ea75417872849d30)]:
+  - @chainlink/coinmarketcap-adapter@2.0.16
+
+## 1.2.48
+
+### Patch Changes
+
+- [#2968](https://github.com/smartcontractkit/external-adapters-js/pull/2968) [`9fc4e5d04`](https://github.com/smartcontractkit/external-adapters-js/commit/9fc4e5d0457379600bcc763c20217dc2331cf941) Thanks [@karen-stepanyan](https://github.com/karen-stepanyan)! - Bumped framework version
+
+- Updated dependencies [[`9fc4e5d04`](https://github.com/smartcontractkit/external-adapters-js/commit/9fc4e5d0457379600bcc763c20217dc2331cf941)]:
+  - @chainlink/coinmarketcap-adapter@2.0.15
+
 ## 1.2.47
 
 ### Patch Changes