+++ packages/sources/cfbenchmarks/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/cfbenchmarks-adapter
 
+## 2.2.1
+
+### Patch Changes
+
+- [#3002](https://github.com/smartcontractkit/external-adapters-js/pull/3002) [`0527b7173`](https://github.com/smartcontractkit/external-adapters-js/commit/0527b7173baaeed5553eb91b714d75587f0dfa64) Thanks [@alecgard](https://github.com/alecgard)! - Fix bug where CFBenchmarks Crypto endpoint would fail if `API_SECONDARY` is explicitly set to false.
+
+## 2.2.0
+
+### Minor Changes
+
+- [#2970](https://github.com/smartcontractkit/external-adapters-js/pull/2970) [`875b24bf6`](https://github.com/smartcontractkit/external-adapters-js/commit/875b24bf6a7e0b4496519b54798d741ab26aa73e) Thanks [@mmcallister-cll](https://github.com/mmcallister-cll)! - Standardize LWBA endpoints
+
+## 2.1.9
+
+### Patch Changes
+
+- [#2968](https://github.com/smartcontractkit/external-adapters-js/pull/2968) [`9fc4e5d04`](https://github.com/smartcontractkit/external-adapters-js/commit/9fc4e5d0457379600bcc763c20217dc2331cf941) Thanks [@karen-stepanyan](https://github.com/karen-stepanyan)! - Bumped framework version
+
 ## 2.1.8
 
 ### Patch Changes