+++ packages/sources/finnhub/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @chainlink/finnhub-adapter
 
+## 2.7.0
+
+### Minor Changes
+
+- [#3016](https://github.com/smartcontractkit/external-adapters-js/pull/3016) [`f5162abff`](https://github.com/smartcontractkit/external-adapters-js/commit/f5162abff871d5733f5e4bbc07c349b0aa41afc8) Thanks [@Subarna-Singh](https://github.com/Subarna-Singh)! - finnhub secondary handling of WHEAT price. Should divide by 100.
+
+- [#3016](https://github.com/smartcontractkit/external-adapters-js/pull/3016) [`f5162abff`](https://github.com/smartcontractkit/external-adapters-js/commit/f5162abff871d5733f5e4bbc07c349b0aa41afc8) Thanks [@Subarna-Singh](https://github.com/Subarna-Singh)! - finnhub secondary handling of WHEAT price. Should divide by 100.
+
+## 2.6.0
+
+### Minor Changes
+
+- [#2999](https://github.com/smartcontractkit/external-adapters-js/pull/2999) [`51304c8e6`](https://github.com/smartcontractkit/external-adapters-js/commit/51304c8e658d9f6f6f41f7731ca50d7ca4cbff8b) Thanks [@boxhock](https://github.com/boxhock)! - Added additional endpoint aliases: "commodities", "forex-quote", "equity-quote" and "commodity-quote"
+
+## 2.5.4
+
+### Patch Changes
+
+- [#2967](https://github.com/smartcontractkit/external-adapters-js/pull/2967) [`a7c807fc6`](https://github.com/smartcontractkit/external-adapters-js/commit/a7c807fc6ce96059c1324381ea75417872849d30) Thanks [@karen-stepanyan](https://github.com/karen-stepanyan)! - Add examples to input parameters
+
+## 2.5.3
+
+### Patch Changes
+
+- [#2968](https://github.com/smartcontractkit/external-adapters-js/pull/2968) [`9fc4e5d04`](https://github.com/smartcontractkit/external-adapters-js/commit/9fc4e5d0457379600bcc763c20217dc2331cf941) Thanks [@karen-stepanyan](https://github.com/karen-stepanyan)! - Bumped framework version
+
 ## 2.5.2
 
 ### Patch Changes