+++ packages/sources/renvm-address-set/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @chainlink/renvm-address-set-adapter
 
+## 1.5.61
+
+### Patch Changes
+
+- Updated dependencies [[`7833c6081`](https://github.com/smartcontractkit/external-adapters-js/commit/7833c60818434709792bf744eb83070ed883efb4)]:
+  - @chainlink/proof-of-reserves-adapter@1.18.14
+
+## 1.5.60
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/proof-of-reserves-adapter@1.18.13
+
+## 1.5.59
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/proof-of-reserves-adapter@1.18.12
+
+## 1.5.58
+
+### Patch Changes
+
+- Updated dependencies [[`9fc4e5d04`](https://github.com/smartcontractkit/external-adapters-js/commit/9fc4e5d0457379600bcc763c20217dc2331cf941)]:
+  - @chainlink/proof-of-reserves-adapter@1.18.11
+
 ## 1.5.57
 
 ### Patch Changes