+++ packages/sources/the-network-firm/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @chainlink/the-network-firm-adapter
 
+## 1.2.0
+
+### Minor Changes
+
+- [#3027](https://github.com/smartcontractkit/external-adapters-js/pull/3027) [`b83080f92`](https://github.com/smartcontractkit/external-adapters-js/commit/b83080f926ab5d43b68d3768fc7f1c2ab5b5c0cc) Thanks [@amit-momin](https://github.com/amit-momin)! - Added GIFT endpoint
+
+- [#2951](https://github.com/smartcontractkit/external-adapters-js/pull/2951) [`0ecddb5d0`](https://github.com/smartcontractkit/external-adapters-js/commit/0ecddb5d02dd7e205c410ea86523a2294e23c2b1) Thanks [@amit-momin](https://github.com/amit-momin)! - Added EURR endpoint
+
+## 1.1.3
+
+### Patch Changes
+
+- [#2967](https://github.com/smartcontractkit/external-adapters-js/pull/2967) [`a7c807fc6`](https://github.com/smartcontractkit/external-adapters-js/commit/a7c807fc6ce96059c1324381ea75417872849d30) Thanks [@karen-stepanyan](https://github.com/karen-stepanyan)! - Add examples to input parameters
+
+## 1.1.2
+
+### Patch Changes
+
+- [#2968](https://github.com/smartcontractkit/external-adapters-js/pull/2968) [`9fc4e5d04`](https://github.com/smartcontractkit/external-adapters-js/commit/9fc4e5d0457379600bcc763c20217dc2331cf941) Thanks [@karen-stepanyan](https://github.com/karen-stepanyan)! - Bumped framework version
+
 ## 1.1.1
 
 ### Patch Changes